requirements.txt
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
flask-wtf>=1.2.1
werkzeug>=3.0.1
wtforms>=3.1.2
python-dotenv>=1.0.1
pytest>=8.1.1
flask-testing>=0.8.1
gunicorn>=21.2.0
email-validator>=2.1.1
